+++ CLI/include/Luau/Coverage.h
